Added a remove file confirmation dialog, and a few bugfixes.

+++ downloads/dropzone-amd-module.js
@@ -241,6 +241,7 @@ Emitter.prototype.hasListeners = function(event){
       dictCancelUpload: "Cancel upload",
       dictCancelUploadConfirmation: "Are you sure you want to cancel this upload?",
       dictRemoveFile: "Remove file",
+      dictRemoveFileConfirmation: null,
       accept: function(file, done) {
         return done();
       },
@@ -346,7 +347,13 @@ Emitter.prototype.hasListeners = function(event){
                 return _this.removeFile(file);
               }
             } else {
-              return _this.removeFile(file);
+              if (_this.options.dictRemoveFileConfirmation) {
+                if (window.confirm(_this.options.dictRemoveFileConfirmation)) {
+                  return _this.removeFile(file);
+                }
+              } else {
+                return _this.removeFile(file);
+              }
             }
           });
           return file.previewElement.appendChild(file._removeLink);
@@ -535,9 +542,7 @@ Emitter.prototype.hasListeners = function(event){
           }
           _this.hiddenFileInput = document.createElement("input");
           _this.hiddenFileInput.setAttribute("type", "file");
-          if (_this.options.uploadMultiple) {
-            _this.hiddenFileInput.setAttribute("multiple", "multiple");
-          }
+          _this.hiddenFileInput.setAttribute("multiple", "multiple");
           if (_this.options.acceptedFiles != null) {
             _this.hiddenFileInput.setAttribute("accept", _this.options.acceptedFiles);
           }
@@ -971,12 +976,15 @@ Emitter.prototype.hasListeners = function(event){
       parallelUploads = this.options.parallelUploads;
       processingLength = this.getUploadingFiles().length;
       i = processingLength;
+      if (processingLength >= parallelUploads) {
+        return;
+      }
       queuedFiles = this.getQueuedFiles();
       if (!(queuedFiles.length > 0)) {
         return;
       }
       if (this.options.uploadMultiple) {
-        return this.processFiles(queuedFiles.slice(0, parallelUploads));
+        return this.processFiles(queuedFiles.slice(0, parallelUploads - processingLength));
       } else {
         while (i < parallelUploads) {
           if (!queuedFiles.length) {
@@ -1055,7 +1063,7 @@ Emitter.prototype.hasListeners = function(event){
     };
 
     Dropzone.prototype.uploadFiles = function(files) {
-      var file, formData, handleError, header, headers, input, inputName, inputType, key, name, progressObj, response, updateProgress, value, xhr, _i, _j, _k, _l, _len, _len1, _len2, _len3, _ref, _ref1, _ref2, _ref3,
+      var file, formData, handleError, headerName, headerValue, headers, input, inputName, inputType, key, progressObj, response, updateProgress, value, xhr, _i, _j, _k, _l, _len, _len1, _len2, _len3, _ref, _ref1, _ref2, _ref3,
         _this = this;
       xhr = new XMLHttpRequest();
       for (_i = 0, _len = files.length; _i < _len; _i++) {
@@ -1148,9 +1156,9 @@ Emitter.prototype.hasListeners = function(event){
       if (this.options.headers) {
         extend(headers, this.options.headers);
       }
-      for (header in headers) {
-        name = headers[header];
-        xhr.setRequestHeader(header, name);
+      for (headerName in headers) {
+        headerValue = headers[headerName];
+        xhr.setRequestHeader(headerName, headerValue);
       }
       formData = new FormData();
       if (this.options.params) {
@@ -1223,7 +1231,7 @@ Emitter.prototype.hasListeners = function(event){
 
   })(Em);
 
-  Dropzone.version = "3.6.1";
+  Dropzone.version = "3.6.2";
 
   Dropzone.options = {};
 
@@ -1251,9 +1259,6 @@ Emitter.prototype.hasListeners = function(event){
 
   Dropzone.discover = function() {
     var checkElements, dropzone, dropzones, _i, _len, _results;
-    if (!Dropzone.autoDiscover) {
-      return;
-    }
     if (document.querySelectorAll) {
       dropzones = document.querySelectorAll(".dropzone");
     } else {
@@ -1506,7 +1511,13 @@ Emitter.prototype.hasListeners = function(event){
     }
   };
 
-  contentLoaded(window, Dropzone.discover);
+  Dropzone._autoDiscoverFunction = function() {
+    if (Dropzone.autoDiscover) {
+      return Dropzone.discover();
+    }
+  };
+
+  contentLoaded(window, Dropzone._autoDiscoverFunction);
 
 }).call(this);
